ajuste preco do produto

+++ Codigo/posto-ipiranga/src/main/java/com/postoipiranga/model/ProductModel.java
@@ -21,6 +21,6 @@ public class ProductModel {
     @Column(nullable = false)
     private Long quantidade;
     @Column(nullable = false)
-    private Long preco;
+    private double preco;
 
 }
